fix(heterovec): fix type mismatch in with_index generated comparison (#12)
SetKey.type_idx was changed from UInt8 to UInt32 for LLVM/SPIR-V
compatibility, but the @generated with_index function still compared
against UInt8 literals. Since Julia's === checks both value and type,
UInt32(1) === UInt8(1) is always false, causing all branches to fall
through to the default (first material). This made every object render
with the same material.

+++ src/heterovec.jl
@@ -490,7 +490,7 @@ The function `f` must not capture variables - pass all data as `args`.
 
     for i in N:-1:1
         result = Expr(:if,
-            :(idx.type_idx === UInt8($i)),
+            :(idx.type_idx === UInt32($i)),
             :(@inbounds f(smv.data[$i][idx.vec_idx], args...)),
             result
         )
